requirements.txt
streamlit>=1.32.0
google-generativeai>=0.7.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
reportlab>=4.1.0
